Extract common testing utils
